Refactoring
Refactoring and cleaning
