+++ source/NyuyuTest/app/src/main/java/com/nilesh/nyuyutest/MainActivity.java
@@ -10,12 +10,12 @@ import android.view.View;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 import com.nilesh.nyuyutest.network.Apicall;
-import com.nilesh.nyuyutest.network.Film;
-import com.nilesh.nyuyutest.network.RecyclerShip;
+import com.nilesh.nyuyutest.model.Film;
+import com.nilesh.nyuyutest.model.RecyclerShip;
 import com.nilesh.nyuyutest.network.StarWarsAPI;
 import com.nilesh.nyuyutest.network.StarWarsClient;
-import com.nilesh.nyuyutest.network.Starships;
-import com.nilesh.nyuyutest.network.finalOutput;
+import com.nilesh.nyuyutest.model.Starships;
+import com.nilesh.nyuyutest.model.finalOutput;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -29,62 +29,47 @@ import rx.schedulers.Schedulers;
 public class MainActivity extends AppCompatActivity {
 
   public static final String TAG = MainActivity.class.getSimpleName();
-
-  private Context mContext;
-
+  private final Map<String, String> mAllFilmURLsMap = new HashMap<>();
+  private final List<Starships> mAllShips = new ArrayList<>();
+  private final List<String> mApiCallLog = new ArrayList<>();
   RecyclerView recyclerView;
   ProgressBar progressBar;
   StarWarsAdapter starWarsAdapter;
   rx.Subscription subscription;
-
+  private Context mContext;
   private StarWarsAdapter mAdapter;
+  private ArrayList<RecyclerShip> mAllRecyclerShips = new ArrayList<>();
 
-  private final Map<String, String> mAllFilmURLsMap = new HashMap<>();
-  private final List<Starships> mAllShips = new ArrayList<>();
-  private final List<String> mApiCallLog = new ArrayList<>();
-  private ArrayList <RecyclerShip> mAllRecyclerShips = new ArrayList<>();
-
-
+  private static Integer compareModel(Starships starship1, Starships starship2) {
+    return starship1.getCostInCredits().compareTo(starship1.getCostInCredits());
+  }
 
   @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_main);
-
 
     mContext = getApplicationContext();
     setupViews();
 
     recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
-
-    //Log.d(TAG, "onCreate: " + savedInstanceState.containsKey("key"));
     if (savedInstanceState != null && savedInstanceState.containsKey("key")) {
       Log.d(TAG, "onCreate: coming ere");
-      recyclerView.setLayoutManager(new LinearLayoutManager(this));
-     // mAllRecyclerShips = savedInstanceState.getParcelableArrayList("key");
-      //categories =  getArguments().getParcelableArrayList("list_categories");
+
       mAllRecyclerShips = savedInstanceState.getParcelableArrayList("key");
       Log.d(TAG, "onCreate: " + mAllRecyclerShips);
       assert mAllRecyclerShips != null;
-      mAdapter = new StarWarsAdapter(mContext, mAllRecyclerShips.subList(0,15));
+      mAdapter = new StarWarsAdapter(mContext, mAllRecyclerShips.subList(0, 15));
       recyclerView.setAdapter(starWarsAdapter);
-    }else {
+    } else {
 
       downloadData();
     }
-
   }
 
   private void setupViews() {
     progressBar = (ProgressBar) findViewById(R.id.rx_progress_bar);
     recyclerView = (RecyclerView) findViewById(R.id.my_recycler_view);
-  }
-
-  private void setupRecyclerView() {
-    recyclerView.setLayoutManager(new LinearLayoutManager(this));
-    //starWarsAdapter = new StarWarsAdapter(new ArrayList<Starships>());
-    //starWarsAdapter = new StarWarsAdapter(mContext, new ArrayList<RecyclerShip>());
-
   }
 
   private void downloadData() {
@@ -93,54 +78,46 @@ public class MainActivity extends AppCompatActivity {
 
     new Apicall(StarWarsClient.getRxApi())
 
-        .getAllStarships(1)
-        .concatMap(new Func1<finalOutput, Observable<Starships>>() {
-          @Override
-          public Observable<Starships> call(finalOutput result) {
-            return Observable.from(result.getShips());
-          }
-        })
+        .getAllStarships(1).concatMap(new Func1<finalOutput, Observable<Starships>>() {
+      @Override public Observable<Starships> call(finalOutput result) {
+        return Observable.from(result.getShips());
+      }
+    })
 
-        .subscribeOn(Schedulers.io())
-        .observeOn(AndroidSchedulers.mainThread())
+        .subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
 
-        .subscribe(new Subscriber<Starships>()  {
-          @Override
-          public void onCompleted() {
-            mAllRecyclerShips = StarshipComparator.populateRecyclerShipList(mAllShips, mAllFilmURLsMap, mAllRecyclerShips);
+        .subscribe(new Subscriber<Starships>() {
+          @Override public void onCompleted() {
+            mAllRecyclerShips =
+                StarshipComparator.populateRecyclerShipList(mAllShips, mAllFilmURLsMap,
+                    mAllRecyclerShips);
             mAllRecyclerShips = StarshipComparator.sortList(mAllRecyclerShips);
             Log.d(TAG, "onCompleted: " + mAllRecyclerShips);
             recyclerView.setVisibility(View.VISIBLE);
             hideProgressBar();
-            mAdapter = new StarWarsAdapter(mContext, mAllRecyclerShips.subList(0,15));
+            mAdapter = new StarWarsAdapter(mContext, mAllRecyclerShips.subList(0, 15));
             recyclerView.setAdapter(mAdapter);
           }
 
-          @Override
-          public void onError(Throwable e) {
+          @Override public void onError(Throwable e) {
             Log.e(TAG, "onError " + e.getMessage());
             e.printStackTrace();
             hideProgressBar();
             Toast.makeText(MainActivity.this, "Cannot download data", Toast.LENGTH_SHORT).show();
           }
 
-
-
           @Override public void onNext(Starships starshipses) {
 
-            for (String film: starshipses.getFilms()) {
-
+            for (String film : starshipses.getFilms()) {
 
               if (!mAllFilmURLsMap.containsKey(film) && !mApiCallLog.contains(film)) {
                 getFilmAndSave(StarWarsClient.getRxApi(), film);
                 mApiCallLog.add(film);
               }
             }
-           mAllShips.add(starshipses);
+            mAllShips.add(starshipses);
             //starWarsAdapter.dataSet.add( mAllShips);
-
           }
-
         });
   }
 
@@ -151,27 +128,20 @@ public class MainActivity extends AppCompatActivity {
         .subscribeOn(Schedulers.io())
         .observeOn(AndroidSchedulers.mainThread())
         .subscribe(new Subscriber<Film>() {
-          @Override
-          public void onCompleted() {
+          @Override public void onCompleted() {
 
           }
 
-          @Override
-          public void onError(Throwable e) {
+          @Override public void onError(Throwable e) {
             Toast.makeText(mContext, e.toString(), Toast.LENGTH_LONG).show();
           }
 
-          @Override
-          public void onNext(Film film) {
+          @Override public void onNext(Film film) {
             //Log.d("MainActivity", film.getTitle());
             mAllFilmURLsMap.put(film.getUrl(), film.getTitle());
             Log.d("MainActivity", mAllFilmURLsMap.toString());
           }
         });
-  }
-
-  private static Integer compareModel(Starships starship1, Starships starship2) {
-    return starship1.getCostInCredits().compareTo(starship1.getCostInCredits());
   }
 
   private void showProgressBar() {
@@ -182,15 +152,15 @@ public class MainActivity extends AppCompatActivity {
     progressBar.setVisibility(View.GONE);
   }
 
-  @Override
-  protected void onDestroy() {
+  @Override protected void onDestroy() {
     if (subscription != null) subscription.unsubscribe();
     super.onDestroy();
   }
 
   protected void onSaveInstanceState(Bundle outState) {
-    if (recyclerView.getAdapter() != null)
+    if (recyclerView.getAdapter() != null) {
       outState.putParcelableArrayList("key", mAllRecyclerShips);
+    }
     super.onSaveInstanceState(outState);
   }
 }
