+++ source/NyuyuTest/app/src/main/java/com/nilesh/nyuyutest/model/RecyclerShip.java
@@ -1,4 +1,4 @@
-package com.nilesh.nyuyutest.network;
+package com.nilesh.nyuyutest.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
