+++ source/NyuyuTest/app/src/main/java/com/nilesh/nyuyutest/model/finalOutput.java
@@ -1,4 +1,4 @@
-package com.nilesh.nyuyutest.network;
+package com.nilesh.nyuyutest.model;
 
 import android.support.annotation.Nullable;
 import com.google.gson.annotations.Expose;
